added router component
