+++ samples/tab-graph-toolkit/csharp/TabGraphToolkit/clientapp/src/components/App.tsx
@@ -1,40 +1,21 @@
-import { Providers, ProviderState } from '@microsoft/mgt';
-import { Agenda, Login } from '@microsoft/mgt-react';
-import React, { useState, useEffect } from 'react';
+import './App.css';
+import {
+    BrowserRouter as Router,
+    Route,
+    Switch
+} from 'react-router-dom';
+import Home from './Home';
 import './App.css';
 
-function useIsSignedIn(): [boolean] {
-    const [isSignedIn, setIsSignedIn] = useState(false);
-
-    useEffect(() => {
-        const updateState = () => {
-            const provider = Providers.globalProvider;
-            setIsSignedIn(provider && provider.state === ProviderState.SignedIn);
-        };
-
-        Providers.onProviderUpdated(updateState);
-        updateState();
-
-        return () => {
-            Providers.removeProviderUpdatedListener(updateState);
-        }
-    }, []);
-
-    return [isSignedIn];
-}
-
 function App() {
-    const [isSignedIn] = useIsSignedIn();
 
     return (
         <div className="App">
-            <header>
-                <Login />
-            </header>
-            <div>
-                {isSignedIn &&
-                    <Agenda />}
-            </div>
+            <Router>
+                <Switch>
+                    <Route exact path='/tab' component={Home}></Route>
+                </Switch>
+            </Router>
         </div>
     );
 }
